+++ geolocator_platform_interface/lib/src/models/models.dart
@@ -1,2 +1,3 @@
 export 'location_options.dart';
+export 'nmea_message.dart';
 export 'position.dart';